Update growGraph.java
The previous version had calculateCenter() which returns centroid's coordinate  (x and y value as an array). It is used to encrypt two characters of the message as a set with that coordinate. It turns out to be major problem when one of the coordinate is not changing after growth/shrink of the triangle, thus, creating limited skipped part of Caesar cipher encryption algorithm. So it is reverted to its original method of encrypting one character at a time.